core/requirements.txt
mock~=4.0.2
fastjsonschema~=2.15.0
boto3~=1.17.8
moto~=1.3.16
parameterized~=0.8.1
setuptools~=60.2.0
jmespath~=0.10.0
cryptography~=38.0.3
pip~=21.3.1
attrs~=22.1.0
toml~=0.10.2
wheel~=0.37.1
Jinja2~=2.11.3
rsa~=4.8
pyasn1~=0.4.8
botocore~=1.20.8
s3transfer~=0.3.4
six~=1.15.0
python-dateutil~=2.8.1
urllib3~=1.26.5
certifi~=2020.12.5
PyJWT~=2.4.0
apispec~=4.3.0